+++ pruebas/01-reading-list/bielmc11/src/components/FiltersComponent.tsx
@@ -1,0 +1,10 @@
+export const FiltersComponent = () => {
+  return (
+    // USAR COMPONENTE DE SHADN/UI
+    // AUNQUE ANTES PUEDE EXPERIMENTEAR CON LOS INPUTS Y EL CSS
+    <div className="inline-block rounded-lg self-start bg-green-500 px-3 py-1 text-sm text-gray-50 tracking-wider">
+                    Filters
+                    <input type="range" />
+    </div>
+  )
+}
