+++ pruebas/01-reading-list/bielmc11/src/components/GeneralList.tsx
@@ -1,8 +1,11 @@
+import { forwardRef } from 'react'
 import { GeneralBooks } from './GeneralBooks'
+import { FiltersComponent } from './FiltersComponent'
 
-export const GeneralList = ({ inputRef }: any) => {
+export const GeneralList = forwardRef<HTMLDivElement>((props, ref) => {
+  GeneralList.displayName = 'GeneralList'
   return (
-    <div className='list bg-gray-200 pt-24' id='listOfBooks' ref={inputRef}>
+    <div className='list bg-gray-200 pt-24' id='listOfBooks' ref={ref}>
 
       <div className='list_wrapper container flex flex-col'>
 
@@ -17,9 +20,10 @@ export const GeneralList = ({ inputRef }: any) => {
           <div className='flex justify-between'>
 
             <p className='text-gray-500 md:text-xl'>Discover our hand-picked collection of the latest and greatest books across a variety of genres.</p>
-            <div className="inline-block rounded-lg self-start bg-green-500 px-3 py-1 text-sm text-gray-50 tracking-wider">
+            {/* <div className="inline-block rounded-lg self-start bg-green-500 px-3 py-1 text-sm text-gray-50 tracking-wider">
                     Filters
-            </div>
+            </div> */}
+            <FiltersComponent/>
           </div>
 
         </div>
@@ -27,4 +31,4 @@ export const GeneralList = ({ inputRef }: any) => {
       </div>
     </div>
   )
-}
+})
